Initial version. Parses entire class file but does not interpret.
This initial version parses an entire .class file, but it doesn't
interpret the code attributes of the various methods. It does however
dump the code in hexadecimal.

The program loads an opcode reference table from opcodes.csv (copy-paste
from Wikipedia). The next stage is to interpret the hexadecimal output
using this reference so it resembles something that is user-readable.